Refactor selectors
