+++ src/applications/vaos/new-appointment/components/CommunityCareProviderSelectionPage/index.jsx
@@ -5,7 +5,7 @@ import SchemaForm from 'platform/forms-system/src/js/components/SchemaForm';
 import FormButtons from '../../../components/FormButtons';
 import { LANGUAGES } from '../../../utils/constants';
 import * as actions from '../../redux/actions';
-import { getFormPageInfo } from '../../../utils/selectors';
+import { getFormPageInfo } from '../../redux/selectors';
 import { scrollAndFocus } from '../../../utils/scrollAndFocus';
 import ProviderSelectionField from './ProviderSelectionField';
