+++ src/applications/vaos/new-appointment/components/ConfirmationPage/ConfirmationRequestInfo.jsx
@@ -2,7 +2,7 @@ import React, { useState } from 'react';
 import moment from 'moment';
 import AlertBox from '@department-of-veterans-affairs/formation-react/AlertBox';
 import AdditionalInfo from '@department-of-veterans-affairs/formation-react/AdditionalInfo';
-import { getTypeOfCare } from '../../../utils/selectors';
+import { getTypeOfCare } from '../../redux/selectors';
 import { FACILITY_TYPES, PURPOSE_TEXT } from '../../../utils/constants';
 import FacilityAddress from '../../../components/FacilityAddress';
