+++ src/applications/vaos/new-appointment/components/DateTimeRequestPage.jsx
@@ -7,7 +7,7 @@ import * as actions from '../redux/actions';
 import { scrollAndFocus } from '../../utils/scrollAndFocus';
 import FormButtons from '../../components/FormButtons';
 import CalendarWidget from './calendar/CalendarWidget';
-import { getFormPageInfo } from '../../utils/selectors';
+import { getFormPageInfo } from '../redux/selectors';
 import { CALENDAR_INDICATOR_TYPES } from '../../utils/constants';
 
 const pageKey = 'requestDateTime';
