+++ src/applications/vaos/new-appointment/components/TypeOfEyeCarePage.jsx
@@ -4,7 +4,7 @@ import { useHistory } from 'react-router-dom';
 import SchemaForm from 'platform/forms-system/src/js/components/SchemaForm';
 import FormButtons from '../../components/FormButtons';
 import * as actions from '../redux/actions';
-import { getFormPageInfo } from '../../utils/selectors';
+import { getFormPageInfo } from '../redux/selectors';
 import { TYPES_OF_EYE_CARE } from '../../utils/constants';
 import { scrollAndFocus } from '../../utils/scrollAndFocus';
