+++ src/applications/vaos/new-appointment/components/VAFacilityPage/index.jsx
@@ -18,7 +18,7 @@ import {
   routeToNextAppointmentPage,
   routeToPreviousAppointmentPage,
 } from '../../redux/actions';
-import { getFacilityPageInfo } from '../../../utils/selectors';
+import { getFacilityPageInfo } from '../../redux/selectors';
 
 import NoVASystems from './NoVASystems';
 import NoValidVAFacilities from './NoValidVAFacilities';
