+++ src/applications/vaos/new-appointment/index.jsx
@@ -11,7 +11,7 @@ import {
   selectUseFlatFacilityPage,
   selectIsCernerOnlyPatient,
   selectUseProviderSelection,
-} from '../utils/selectors';
+} from '../redux/selectors';
 import newAppointmentReducer from './redux/reducer';
 import FormLayout from './components/FormLayout';
 import TypeOfCarePage from './components/TypeOfCarePage';
