+++ src/applications/vaos/new-appointment/redux/selectors.js
@@ -1,88 +1,30 @@
-import moment from 'moment';
-import { createSelector } from 'reselect';
-import { toggleValues } from 'platform/site-wide/feature-toggles/selectors';
 import {
-  selectPatientFacilities,
   selectVAPResidentialAddress,
   selectCernerAppointmentsFacilities,
 } from 'platform/user/selectors';
-import { titleCase } from './formatters';
+
 import {
   getTimezoneBySystemId,
   getTimezoneDescBySystemId,
-  getTimezoneAbbrBySystemId,
-} from './timezone';
-import { isEligible } from '../new-appointment/redux/helpers/eligibility';
+} from '../../utils/timezone';
 import {
   FACILITY_TYPES,
   TYPES_OF_CARE,
   TYPES_OF_SLEEP_CARE,
   TYPES_OF_EYE_CARE,
   FETCH_STATUS,
-  APPOINTMENT_STATUS,
   AUDIOLOGY_TYPES_OF_CARE,
-} from './constants';
-import { getSiteIdFromOrganization } from '../services/organization';
+} from '../../utils/constants';
+import { getSiteIdFromOrganization } from '../../services/organization';
 import {
   getParentOfLocation,
   getSiteIdFromFakeFHIRId,
-} from '../services/location';
+} from '../../services/location';
+import { isEligible } from './helpers/eligibility';
 import {
-  getVideoAppointmentLocation,
-  getVAAppointmentLocationId,
-  isVideoAppointment,
-  isUpcomingAppointmentOrRequest,
-  isValidPastAppointment,
-  sortByDateDescending,
-  sortUpcoming,
-  getVARFacilityId,
-} from '../services/appointment';
-
-export const selectIsCernerOnlyPatient = state =>
-  !!selectPatientFacilities(state)?.every(
-    f => f.isCerner && f.usesCernerAppointments,
-  );
-
-export const selectIsCernerPatient = state =>
-  selectPatientFacilities(state)?.some(
-    f => f.isCerner && f.usesCernerAppointments,
-  );
-
-export const selectIsRegisteredToSacramentoVA = state =>
-  selectPatientFacilities(state)?.some(f => f.facilityId === '612');
-
-export const vaosApplication = state => toggleValues(state).vaOnlineScheduling;
-export const vaosCancel = state => toggleValues(state).vaOnlineSchedulingCancel;
-export const vaosRequests = state =>
-  toggleValues(state).vaOnlineSchedulingRequests;
-export const vaosCommunityCare = state =>
-  toggleValues(state).vaOnlineSchedulingCommunityCare;
-export const vaosDirectScheduling = state =>
-  toggleValues(state).vaOnlineSchedulingDirect;
-export const vaosPastAppts = state =>
-  toggleValues(state).vaOnlineSchedulingPast;
-export const vaosVSPAppointmentNew = state =>
-  toggleValues(state).vaOnlineSchedulingVspAppointmentNew;
-export const vaosExpressCare = state =>
-  toggleValues(state).vaOnlineSchedulingExpressCare;
-export const vaosExpressCareNew = state =>
-  toggleValues(state).vaOnlineSchedulingExpressCareNew;
-export const selectFeatureToggleLoading = state => toggleValues(state).loading;
-const vaosFlatFacilityPage = state =>
-  toggleValues(state).vaOnlineSchedulingFlatFacilityPage;
-const vaosFlatFacilityPageSacramento = state =>
-  toggleValues(state).vaOnlineSchedulingFlatFacilityPageSacramento;
-export const selectUseFlatFacilityPage = state =>
-  vaosFlatFacilityPage(state) &&
-  !selectIsCernerPatient(state) &&
-  (!selectIsRegisteredToSacramentoVA(state) ||
-    vaosFlatFacilityPageSacramento(state));
-
-const vaosProviderSelection = state =>
-  toggleValues(state).vaOnlineSchedulingProviderSelection;
-export const selectUseProviderSelection = state =>
-  vaosProviderSelection(state) &&
-  !!selectVAPResidentialAddress(state)?.addressLine1;
+  selectUseFlatFacilityPage,
+  selectIsCernerOnlyPatient,
+} from '../../redux/selectors';
 
 export function getNewAppointment(state) {
   return state.newAppointment;
@@ -439,50 +381,6 @@ export function getClinicPageInfo(state, pageKey) {
   };
 }
 
-export function getCancelInfo(state) {
-  const {
-    appointmentToCancel,
-    showCancelModal,
-    cancelAppointmentStatus,
-    cancelAppointmentStatusVaos400,
-    facilityData,
-  } = state.appointments;
-
-  const isVideo = appointmentToCancel
-    ? isVideoAppointment(appointmentToCancel)
-    : false;
-
-  let facility = null;
-  if (appointmentToCancel?.status === APPOINTMENT_STATUS.booked && !isVideo) {
-    // Confirmed in person VA appts
-    const locationId = getVAAppointmentLocationId(appointmentToCancel);
-    facility = facilityData[locationId];
-  } else if (appointmentToCancel?.facility) {
-    // Requests
-    facility = facilityData[`var${appointmentToCancel.facility.facilityCode}`];
-  } else if (isVideo) {
-    // Video visits
-    const locationId = getVideoAppointmentLocation(appointmentToCancel);
-    facility = facilityData[locationId];
-  }
-  let isCerner = null;
-  if (appointmentToCancel) {
-    const facilityId = getVARFacilityId(appointmentToCancel);
-    isCerner = selectCernerAppointmentsFacilities(state)?.some(cernerSite =>
-      facilityId?.startsWith(cernerSite.facilityId),
-    );
-  }
-
-  return {
-    facility,
-    appointmentToCancel,
-    showCancelModal,
-    cancelAppointmentStatus,
-    cancelAppointmentStatusVaos400,
-    isCerner,
-  };
-}
-
 export function getChosenVACityState(state) {
   const schema =
     state.newAppointment.pages.ccPreferences?.properties.communityCareSystemId;
@@ -496,261 +394,4 @@ export function getChosenVACityState(state) {
   }
 
   return null;
-}
-
-export const selectIsWelcomeModalDismissed = state =>
-  state.announcements.dismissed.some(
-    announcement => announcement === 'welcome-to-new-vaos',
-  );
-
-export const selectSystemIds = state =>
-  selectPatientFacilities(state)?.map(f => f.facilityId) || null;
-
-export const selectExpressCareRequests = createSelector(
-  state => state.appointments.pending,
-  pending =>
-    pending?.filter(appt => appt.vaos.isExpressCare).sort(sortByDateDescending),
-);
-
-export function selectFutureStatus(state) {
-  const { pendingStatus, confirmedStatus } = state.appointments;
-  if (
-    pendingStatus === FETCH_STATUS.failed ||
-    confirmedStatus === FETCH_STATUS.failed
-  ) {
-    return FETCH_STATUS.failed;
-  }
-
-  if (
-    pendingStatus === FETCH_STATUS.loading ||
-    confirmedStatus === FETCH_STATUS.loading
-  ) {
-    return FETCH_STATUS.loading;
-  }
-
-  if (
-    pendingStatus === FETCH_STATUS.succeeded &&
-    confirmedStatus === FETCH_STATUS.succeeded
-  ) {
-    return FETCH_STATUS.succeeded;
-  }
-
-  return FETCH_STATUS.notStarted;
-}
-
-export const selectFutureAppointments = createSelector(
-  vaosExpressCare,
-  state => state.appointments.pending,
-  state => state.appointments.confirmed,
-  (showExpressCare, pending, confirmed) => {
-    if (!confirmed || !pending) {
-      return null;
-    }
-
-    return confirmed
-      .concat(...pending)
-      .filter(appt => !showExpressCare || !appt.vaos.isExpressCare)
-      .filter(isUpcomingAppointmentOrRequest)
-      .sort(sortUpcoming);
-  },
-);
-
-export const selectPastAppointments = createSelector(
-  state => state.appointments.past,
-  past => {
-    return past?.filter(isValidPastAppointment).sort(sortByDateDescending);
-  },
-);
-
-export function selectExpressCareNewRequest(state) {
-  return state.expressCare.newRequest;
-}
-
-export function selectExpressCareFormData(state) {
-  return selectExpressCareNewRequest(state).data;
-}
-
-export function selectExpressCareFacilities(state) {
-  return state.appointments.expressCareFacilities;
-}
-
-/*
- * Selects any EC windows that we're in at the current (or provided) time
- */
-export function selectActiveExpressCareWindows(state, nowMoment) {
-  const now = nowMoment || moment();
-  return selectExpressCareFacilities(state)
-    ?.map(({ days, facilityId }) => {
-      const siteId = facilityId.substring(0, 3);
-      const { timezone } = getTimezoneBySystemId(siteId);
-      const timezoneAbbreviation = getTimezoneAbbrBySystemId(siteId);
-      const nowFacilityTime = now.clone().tz(timezone);
-      const currentDayOfWeek = nowFacilityTime.format('dddd').toUpperCase();
-      const activeDay = days.find(day => day.day === currentDayOfWeek);
-
-      if (!activeDay) {
-        return null;
-      }
-
-      const start = moment.tz(
-        `${nowFacilityTime.format('YYYY-MM-DD')}T${activeDay.startTime}:00`,
-        timezone,
-      );
-      const end = moment.tz(
-        `${nowFacilityTime.format('YYYY-MM-DD')}T${activeDay.endTime}:00`,
-        timezone,
-      );
-
-      if (!now.isBetween(start, end)) {
-        return null;
-      }
-
-      return {
-        facilityId,
-        siteId,
-        timezone,
-        timezoneAbbreviation,
-        start,
-        end,
-      };
-    })
-    .filter(win => !!win);
-}
-
-/*
- * Gets the formatted hours string of the current window, chosen based on the
- * provided time.
- *
- * Note: we're picking the first active window, there could be more than one
- */
-export function selectLocalExpressCareWindowString(state, nowMoment) {
-  const current = selectActiveExpressCareWindows(state, nowMoment);
-
-  if (!current?.length) {
-    return null;
-  }
-
-  return `${current[0].start.format('h:mm a')} to ${current[0].end.format(
-    'h:mm a',
-  )} ${current[0].timezoneAbbreviation}`;
-}
-
-/*
- * Gets the facility info for the current window, chosen based on the
- * provided time.
- *
- * Note: we're picking the first active window, there could be more than one
- */
-export function selectActiveExpressCareFacility(state, nowMoment) {
-  const current = selectActiveExpressCareWindows(state, nowMoment);
-
-  if (!current?.length) {
-    return null;
-  }
-
-  return {
-    facilityId: current[0].facilityId,
-    siteId: current[0].facilityId.substring(0, 3),
-  };
-}
-
-function getFormattedTime(time) {
-  return moment(`${moment().format('YYYY-MM-DD')}T${time}:00`).format('h:mm a');
-}
-
-function getWindowString(window, timezoneAbbreviation, isToday) {
-  return `${isToday ? 'today' : titleCase(window.day)} from ${getFormattedTime(
-    window.startTime,
-  )} to ${getFormattedTime(window.endTime)} ${timezoneAbbreviation}`;
-}
-
-/**
- * Returns next schedulable window.  If today is schedulable and current time is before window,
- * return today's window.  Otherwise, return the next schedulable day's window
- */
-export function selectNextAvailableExpressCareWindowString(state, nowMoment) {
-  const supportedFacilities = selectExpressCareFacilities(state);
-  if (!supportedFacilities?.length) {
-    return null;
-  }
-
-  const facility = supportedFacilities[0];
-  const siteId = facility.facilityId.substring(0, 3);
-  const { timezone } = getTimezoneBySystemId(siteId);
-  const timezoneAbbreviation = getTimezoneAbbrBySystemId(siteId);
-  const nowFacilityTime = nowMoment.clone().tz(timezone);
-  const dayOfWeek = nowFacilityTime.format('dddd').toUpperCase();
-  const todayDayOfWeekIndex = Number(nowFacilityTime.format('d'));
-  const todaysWindow = facility.days.find(d => d.day === dayOfWeek);
-
-  // Sort schedulable days after today so we can easily find the next
-  // day if needed
-  const schedulableDaysAfterToday = [
-    ...facility.days.filter(day => day.dayOfWeekIndex > todayDayOfWeekIndex),
-    ...facility.days.filter(day => day.dayOfWeekIndex < todayDayOfWeekIndex),
-  ];
-
-  if (todaysWindow) {
-    const start = moment.tz(
-      `${nowFacilityTime.format('YYYY-MM-DD')}T${todaysWindow.startTime}:00`,
-      timezone,
-    );
-    if (nowMoment.isBefore(start)) {
-      // If today is schedulable and we are before the window, return today's window
-      return getWindowString(todaysWindow, timezoneAbbreviation, true);
-    } else {
-      // In the rare case the today is the only schedulable day and they are past the window,
-      // return today's window and specify "next week". Otherwise, return the next schedulable day
-      if (!schedulableDaysAfterToday.length) {
-        return `next ${getWindowString(
-          todaysWindow,
-          timezoneAbbreviation,
-          false,
-        )}`;
-      }
-      return getWindowString(
-        schedulableDaysAfterToday[0],
-        timezoneAbbreviation,
-        false,
-      );
-    }
-  } else {
-    // If today isn't schedulable, return the next day that is
-    return getWindowString(
-      schedulableDaysAfterToday[0],
-      timezoneAbbreviation,
-      false,
-    );
-  }
-}
-
-export function selectExpressCare(state) {
-  const expressCare = state.expressCare;
-  const activeWindows = selectActiveExpressCareWindows(state);
-  return {
-    ...expressCare,
-    activeWindows,
-    localWindowString: selectLocalExpressCareWindowString(state),
-    localNextAvailableString: selectNextAvailableExpressCareWindowString(
-      state,
-      moment(),
-    ),
-    allowRequests: !!activeWindows?.length,
-    enabled: vaosExpressCare(state),
-    useNewFlow: vaosExpressCareNew(state),
-    hasWindow: !!selectExpressCareFacilities(state)?.length,
-    hasRequests:
-      vaosExpressCare(state) &&
-      state.appointments.pending?.some(appt => appt.vaos.isExpressCare),
-    windowsStatus: state.appointments.expressCareWindowsStatus,
-  };
-}
-
-export function getExpressCareFormPageInfo(state, pageKey) {
-  const newRequest = selectExpressCareNewRequest(state);
-  return {
-    schema: newRequest.pages[pageKey],
-    data: newRequest.data,
-    pageChangeInProgress: newRequest.pageChangeInProgress,
-  };
 }
